% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_dataset}
\alias{combined_dataset}
\title{Combined Stock-Recruitment dataset}
\format{
A data frame with columns:
\describe{
  \item{Year}{Year of observation}
  \item{SSB}{Spawning stock biomass}
  \item{Recruitment}{Recruitment values}
  \item{species}{Species name}
}
}
\source{
Generated from stockassessment.org HTML files
}
\usage{
combined_dataset
}
\description{
A dataset containing SSB and Recruitment time series data for different
species of Norwegian fish stock.
}
\keyword{datasets}
