% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{hockey_model}
\alias{hockey_model}
\title{Hockey Stick model.}
\usage{
hockey_model(data, a_start = NULL, b_start = NULL)
}
\arguments{
\item{data}{Takes in the dataset}

\item{a_start}{Lets user choose a starting value for a}

\item{b_start}{Lets user choose a starting value for b}
}
\value{
returns a nlsLM model
}
\description{
Fits a Hockey Stick model using nls.
}
\examples{
data <- tibble(
SSB = c(1000,2000,3000,4000,5000),
Recruitment = c(1500,2500,3500,4500,5500))

fit <- hockey_model(data)
coef(fit)

pred <- predict_hockey(fit,data)
head(pred)
}
