% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_cleaning_merging_data.R
\name{import_species_data}
\alias{import_species_data}
\title{Import stock assessment html file}
\usage{
import_species_data(file_path, species_name)
}
\arguments{
\item{file_path}{Takes in file path of an html file}

\item{species_name}{Takes a species name chosen by user (string)}
}
\value{
returns a cleaned html file
}
\description{
Imports and cleans a stock assessment html file
}
