% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_cleaning_merging_data.R
\name{merge_species_df}
\alias{merge_species_df}
\title{Import and merge stock assessment datasets}
\usage{
merge_species_df(file_path_list, species_names_list)
}
\arguments{
\item{file_path_list}{Takes in a list of file paths}

\item{species_names_list}{Takes in a list of species names in order of file list}
}
\value{
returns a merged dataset
}
\description{
Takes files and runs them through import_species_data and then merges them
}
