% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_bh}
\alias{predict_bh}
\title{Beverton & Holt prediction.}
\usage{
predict_bh(model, data)
}
\arguments{
\item{model}{Takes in the Ricker model}

\item{data}{Takes in the dataset}
}
\value{
returns a predicted dataset
}
\description{
Predicts new data to plot the Beverton & Holt model.
}
\examples{
data <- data.frame(
SSB = c(1000,2000,3000,4000,5000),
Recruitment = c(1500,2500,3500,4500,5500))

fit <- bh_model(data)
coef(fit)

pred <- predict_bh(fit,data)
head(pred)
}
