% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_hockey}
\alias{predict_hockey}
\title{Hockey Stick prediction.}
\usage{
predict_hockey(model, data)
}
\arguments{
\item{model}{Takes in the Hockey Stick model}

\item{data}{Takes in the dataset}
}
\value{
returns a predicted dataset
}
\description{
Predicts new data to plot the Hockey Stick model.
}
\examples{
data <- data.frame(
SSB = c(1000,2000,3000,4000,5000),
Recruitment = c(1500,2500,3500,4500,5500))

fit <- hockey_model(data)
coef(fit)

pred <- predict_hockey(fit,data)
head(pred)
}
