% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_ricker}
\alias{predict_ricker}
\title{Ricker prediction.}
\usage{
predict_ricker(model, data)
}
\arguments{
\item{model}{Takes in the Ricker model}

\item{data}{Takes in the dataset}
}
\value{
returns a predicted dataset
}
\description{
Predicts new data to plot the Ricker model.
}
\examples{
data <- tibble(
SSB = c(1000,2000,3000,4000,5000),
Recruitment = c(1500,2500,3500,4500,5500))

fit <- ricker_model(data)
coef(fit)

pred <- predict_ricker(fit,data)
head(pred)
}
