% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ricker_model}
\alias{ricker_model}
\title{Ricker model.}
\usage{
ricker_model(data, a_start = NULL, b_start = NULL)
}
\arguments{
\item{data, a_start, b_start}{model takes in the Ricker model.
data takes in the dataset
a_start lets user choose a starting value for a
b_start lets user choose a starting value for b}
}
\value{
returns a nlsLM model
}
\description{
Fits a Ricker model using nls
}
\examples{
data <- tibble(
SSB = c(1000,2000,3000,4000,5000),
Recruitment = c(1500,2500,3500,4500,5500))

fit <- ricker_model(data)
coef(fit)

pred <- predict_ricker(fit,data)
head(pred)
}
