% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_cleaning_merging_data.R
\name{merge_species_df}
\alias{merge_species_df}
\title{Merges already cleaned datasets}
\usage{
merge_species_df(file_path_list, species_names_list)
}
\arguments{
\item{file_path, species_name}{file_path takes in one file path of an html file
species_name takes a species name completely chosen by user (string)}
}
\value{
returns a cleaned html file
}
\description{
Merges already cleaned datasets
}
